refactor(api): extract JSON-RPC types to zeroclaw-api::jsonrpc
Move JsonRpcRequest, JsonRpcResponse, JsonRpcNotification, JsonRpcError,
error_codes, ACP_PROTOCOL_VERSION, and RpcOutbound from
zeroclaw-channels::orchestrator::acp_server into zeroclaw-api::jsonrpc.

This allows the new runtime RPC layer (#6837) to share wire types with
the ACP server without depending on zeroclaw-channels.

- RpcOutbound::new() is now pub (was crate-private)
- RpcOutbound::send_raw() added for direct line writes
- RpcOutbound::pending_count() replaces test-only pending_count_for_test()
- acp_server.rs write_json() updated to use send_raw()
- acp_channel.rs tests updated to use public API directly

Part of #6837.
